rm_oms_apis/pyproject.toml
[tool.poetry]
name = "rm-oms-apis"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pydantic = "^2.7.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
bcrypt = "^4.1.3"
jose = "^1.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
